+++ src/editor/ha-form.ts
@@ -18,7 +18,8 @@ export type HaFormSchema =
   | HaFormTimeSchema
   | HaFormSelector
   | HaFormGridSchema
-  | HaFormExpandableSchema;
+  | HaFormExpandableSchema
+  | HaFormOptionalActionsSchema;
 
 export interface HaFormBaseSchema {
   name: string;
@@ -49,6 +50,12 @@ export interface HaFormExpandableSchema extends HaFormBaseSchema {
   iconPath?: string;
   expanded?: boolean;
   headingLevel?: 1 | 2 | 3 | 4 | 5 | 6;
+  schema: readonly HaFormSchema[];
+}
+
+export interface HaFormOptionalActionsSchema extends HaFormBaseSchema {
+  type: "optional_actions";
+  flatten?: boolean;
   schema: readonly HaFormSchema[];
 }
